database/requirements.txt
Faker~=25.8.0
psycopg~=3.1.19
python-dotenv
